requirements/dev.txt
-r common.txt

coverage==7.6.1
pyodbc==5.1.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0

